Updated styling of SVG line chart
The styling added surrounds the font of the labels (in order to match the other line chart created with Canvas). The labels were also rotated to the same degree as the chart created with Chartist, in order to make them as identical as possible. The thickness of the line has also been reduced to match the other diagram.